Rename all the imports as a result of the modularization in the migration module
