+++ kin-devplatform-sdk/build.gradle
@@ -19,7 +19,7 @@ android {
 dependencies {
 
     implementation project(':kin-devplatform-core')
-    implementation 'com.github.kinecosystem:kin-migration-module-android:7d4ef2119f'
+    implementation 'com.github.kinecosystem:kin-migration-module-android:1.0.4'
     implementation 'com.github.kinecosystem.kin-ecosystem-android-sdk:recovery:d73ef56'
     api 'com.github.kinecosystem:kin-utils-android:1.1'
     implementation devDependencies.appCompat_v7
